refactor(aggregator): move builders function to `support` submodule
